refactor: the initialization for fakeimage
